Fixed the problem with buffers getting blocked and never waking up.  Ran into a few issues:  there was a segfault during some testing with high workers (500+) but I can't reproduce it; status printing really needs cleaned up; getting an error when trying to decompress a buffer as part of a restoration (pretty sure it's because another thread already restored it, but we'll see...).  All in all I'm happy for now because this bug was a huge waste of time, easily 40-50 hours debugging it.

+++ src/list.h
@@ -111,7 +111,7 @@ int list__update_ref(List *list, int delta);
 int list__search(List *list, Buffer **buf, bufferid_t id);
 int list__acquire_write_lock(List *list);
 int list__release_write_lock(List *list);
-uint32_t list__sweep(List *list, uint8_t sweep_goal);
+uint64_t list__sweep(List *list, uint8_t sweep_goal);
 int list__balance(List *list, uint32_t ratio);
 int list__destroy(List *list);
 void list__compressor_start(Compressor *comp);
